requirements.txt
Flask==1.1.2
gunicorn==20.0.4
opencensus-ext-azure==1.0.4
